Animation function and new art asset
Update the animation function. When player moves, the animation is playing animator files. The code use the detection of GetAxisRaw to make boolean variable that control the play of animation. This feature has a bug that the animation cannot be played correctly when player doing x and y axis motion at the same time. The art asset are draw by Jingyu.